+++ CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 
 ### Changed
 
-- Remove unused macros from `libs/common/macros.h`
+- Remove unused macros from `libs/common/macros.h`.
+- Use static event `m_evt_stopped` in `apps/examples/workers`.
 
 ## v0.12.0 - 23-July-2025